AD-15 feat: insight 페이지 타입 정의 분리
