+++ src/pages/insight/insightMetrics.ts
@@ -1,19 +1,6 @@
 // 인사이트 대시보드: 지표 계산 및 일자별 스냅샷 유틸리티
 import type { Task } from "../../types/task";
-
-// 내부적으로 선택 타임스탬프를 허용하기 위한 보조 타입
-interface TaskWithTimestamps extends Task {
-  createdAt?: string;
-  updatedAt?: string;
-  qaApprovedAt?: string;
-}
-
-export interface InsightMetricsSnapshot {
-  dateKey: string; // KST YYYYMMDD
-  completionRate: number; // 0~100
-  qaDoneCount: number; // DONE 개수
-  lastProgressAt?: string; // ISO 문자열
-}
+import type { InsightMetricsSnapshot, TaskWithTimestamps } from "./insight";
 
 const STORAGE_PREFIX = "insight:metrics";
